chore: restructuring temporarily to decrease complexity while ironing out the last details
