+++ src/wayland/input_method/mod.rs
@@ -80,8 +80,7 @@ use crate::{
     utils::{Logical, Rectangle},
 };
 
-pub use input_method_handle::{InputMethodHandle, InputMethodUserData};
-pub use input_method_keyboard_grab::{InputMethodKeyboardGrab, InputMethodKeyboardUserData};
+pub use input_method_handle::{InputMethodHandle, InputMethodKeyboardUserData, InputMethodUserData};
 pub use input_method_popup_surface::InputMethodPopupSurfaceUserData;
 
 use super::text_input::TextInputHandle;
@@ -92,7 +91,6 @@ const MANAGER_VERSION: u32 = 1;
 pub const INPUT_POPUP_SURFACE_ROLE: &str = "zwp_input_popup_surface_v2";
 
 mod input_method_handle;
-mod input_method_keyboard_grab;
 mod input_method_popup_surface;
 pub use input_method_popup_surface::{PopupParent, PopupSurface};
 
@@ -227,7 +225,7 @@ where
                         dismiss_popup: D::dismiss_popup,
                     },
                 );
-                handle.add_instance(&instance, client, dh);
+                handle.add_instance(&instance, client, dh, text_input_handle.clone());
             }
             zwp_input_method_manager_v2::Request::Destroy => {
                 // Nothing to do
